requirements.txt
python-dotenv
imapclient
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
numpy
apscheduler

pip
rsa
pyasn1
pytz
setuptools
pyasn1-modules
cachetools
certifi
urllib3
requests
tzlocal
pyparsing
httplib2
oauthlib
idna
six
uritemplate
schedule
pyinstaller

pyasn1_modules
packaging
altgraph
pefile

pyasn1_modules
packaging
altgraph
pefile

